Aggregate branch hits between tags for better partial detection from multiple uploads
