+++ services/analyse/src/Query/FileCoverageQuery.php
@@ -39,7 +39,7 @@ class FileCoverageQuery extends AbstractLineCoverageQuery
                 2
             ) as coveragePercentage
         FROM
-            lineCoverageWithState
+            lines
         GROUP BY
             fileName
         ORDER BY
@@ -52,38 +52,6 @@ class FileCoverageQuery extends AbstractLineCoverageQuery
             )
             ASC
         {$limit}
-        SQL;
-    }
-
-    public function getNamedQueries(string $table, Upload $upload, ?QueryParameterBag $parameterBag = null): string
-    {
-        $parent = parent::getNamedQueries($table, $upload, $parameterBag);
-
-        $covered = LineState::COVERED->value;
-        $partial = LineState::PARTIAL->value;
-        $uncovered = LineState::UNCOVERED->value;
-
-        return <<<SQL
-        {$parent},
-        lineCoverageWithState AS (
-            SELECT
-                fileName,
-                lineNumber,
-                IF(
-                    SUM(hits) = 0,
-                    "{$uncovered}",
-                    IF (
-                        MIN(isPartiallyHit) = 1,
-                        "{$partial}",
-                        "{$covered}"
-                    )
-                ) as state
-            FROM
-                lineCoverage
-            GROUP BY
-                fileName,
-                lineNumber
-        )
         SQL;
     }
