+++ services/analyse/src/Query/LineCoverageQuery.php
@@ -8,7 +8,6 @@ use App\Query\Result\MultiLineCoverageQueryResult;
 use App\Query\Trait\ScopeAwareTrait;
 use Google\Cloud\BigQuery\QueryResults;
 use Google\Cloud\Core\Exception\GoogleException;
-use Packages\Models\Enum\LineState;
 use Packages\Models\Model\Upload;
 
 class LineCoverageQuery extends AbstractLineCoverageQuery
@@ -17,29 +16,12 @@ class LineCoverageQuery extends AbstractLineCoverageQuery
 
     public function getQuery(string $table, Upload $upload, ?QueryParameterBag $parameterBag = null): string
     {
-        $covered = LineState::COVERED->value;
-        $partial = LineState::PARTIAL->value;
-        $uncovered = LineState::UNCOVERED->value;
-
         return <<<SQL
         {$this->getNamedQueries($table, $upload, $parameterBag)}
         SELECT
-            fileName,
-            lineNumber,
-            IF(
-                SUM(hits) = 0,
-                "{$uncovered}",
-                IF (
-                    MIN(isPartiallyHit) = 1,
-                    "{$partial}",
-                    "{$covered}"
-                )
-            ) as state
+            *
         FROM
-            lineCoverage
-        GROUP BY
-            fileName,
-            lineNumber
+            lines
         SQL;
     }
