+++ services/analyse/tests/Query/FileCoverageQueryTest.php
@@ -36,22 +36,26 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                         WHERE
                             key = "lineHits"
                     ) AS hits,
-                    IF (
-                        type = "BRANCH",
-                        (
-                            SELECT
-                              IF (
-                                value <> '',
-                                CAST(value AS int),
-                                0
-                              )
-                            FROM
-                              UNNEST(metadata)
-                            WHERE
-                              KEY = "partial"
-                        ),
-                        0
-                    ) AS isPartiallyHit
+                    ARRAY(
+                        SELECT
+                            SUM(CAST(branchHits AS INT64))
+                        FROM
+                            UNNEST(
+                                JSON_VALUE_ARRAY(
+                                    (
+                                        SELECT
+                                            value
+                                        FROM
+                                            UNNEST(metadata)
+                                        WHERE
+                                            KEY = "branchHits"
+                                    )
+                                ) 
+                            ) AS branchHits WITH OFFSET AS branchIndex
+                        GROUP BY
+                            branchIndex,
+                            branchHits
+                    ) as branchHits
                 FROM
                     `mock-table`
                 WHERE
@@ -66,21 +70,28 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                     lineNumber IN (10,11,12)
                 ))
             ),
-            lineCoverage AS (
+            branchingLines AS (
                 SELECT
                     fileName,
                     lineNumber,
-                    tag,
                     SUM(hits) as hits,
-                    MIN(isPartiallyHit) as isPartiallyHit
-                FROM
-                    unnested
-                GROUP BY
+                    branchIndex,
+                    SUM(branchHit) > 0 as isBranchedLineHit
+                FROM 
+                    unnested,
+                    UNNEST(
+                        IF(
+                            ARRAY_LENGTH(branchHits) = 0,
+                            [hits],
+                            branchHits    
+                        )
+                    ) AS branchHit WITH OFFSET AS branchIndex
+                GROUP BY 
                     fileName,
                     lineNumber,
-                    tag
+                    branchIndex
             ),
-            lineCoverageWithState AS (
+            lines AS (
                 SELECT
                     fileName,
                     lineNumber,
@@ -88,13 +99,13 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                         SUM(hits) = 0,
                         "uncovered",
                         IF (
-                            MIN(isPartiallyHit) = 1,
+                            MIN(CAST(isBranchedLineHit AS INT64)) = 0,
                             "partial",
                             "covered"
                         )
                     ) as state
                 FROM
-                    lineCoverage
+                    branchingLines
                 GROUP BY
                     fileName,
                     lineNumber
@@ -113,7 +124,7 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                     2
                 ) as coveragePercentage
             FROM
-                lineCoverageWithState
+                lines
             GROUP BY
                 fileName
             ORDER BY
@@ -143,22 +154,26 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                         WHERE
                             key = "lineHits"
                     ) AS hits,
-                    IF (
-                        type = "BRANCH",
-                        (
-                            SELECT
-                              IF (
-                                value <> '',
-                                CAST(value AS int),
-                                0
-                              )
-                            FROM
-                              UNNEST(metadata)
-                            WHERE
-                              KEY = "partial"
-                        ),
-                        0
-                    ) AS isPartiallyHit
+                    ARRAY(
+                        SELECT
+                            SUM(CAST(branchHits AS INT64))
+                        FROM
+                            UNNEST(
+                                JSON_VALUE_ARRAY(
+                                    (
+                                        SELECT
+                                            value
+                                        FROM
+                                            UNNEST(metadata)
+                                        WHERE
+                                            KEY = "branchHits"
+                                    )
+                                ) 
+                            ) AS branchHits WITH OFFSET AS branchIndex
+                        GROUP BY
+                            branchIndex,
+                            branchHits
+                    ) as branchHits
                 FROM
                     `mock-table`
                 WHERE
@@ -173,21 +188,28 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                     lineNumber IN (10,11,12)
                 ))
             ),
-            lineCoverage AS (
+            branchingLines AS (
                 SELECT
                     fileName,
                     lineNumber,
-                    tag,
                     SUM(hits) as hits,
-                    MIN(isPartiallyHit) as isPartiallyHit
-                FROM
-                    unnested
-                GROUP BY
+                    branchIndex,
+                    SUM(branchHit) > 0 as isBranchedLineHit
+                FROM 
+                    unnested,
+                    UNNEST(
+                        IF(
+                            ARRAY_LENGTH(branchHits) = 0,
+                            [hits],
+                            branchHits    
+                        )
+                    ) AS branchHit WITH OFFSET AS branchIndex
+                GROUP BY 
                     fileName,
                     lineNumber,
-                    tag
+                    branchIndex
             ),
-            lineCoverageWithState AS (
+            lines AS (
                 SELECT
                     fileName,
                     lineNumber,
@@ -195,13 +217,13 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                         SUM(hits) = 0,
                         "uncovered",
                         IF (
-                            MIN(isPartiallyHit) = 1,
+                            MIN(CAST(isBranchedLineHit AS INT64)) = 0,
                             "partial",
                             "covered"
                         )
                     ) as state
                 FROM
-                    lineCoverage
+                    branchingLines
                 GROUP BY
                     fileName,
                     lineNumber
@@ -220,7 +242,7 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                     2
                 ) as coveragePercentage
             FROM
-                lineCoverageWithState
+                lines
             GROUP BY
                 fileName
             ORDER BY
@@ -250,22 +272,26 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                         WHERE
                             key = "lineHits"
                     ) AS hits,
-                    IF (
-                        type = "BRANCH",
-                        (
-                            SELECT
-                              IF (
-                                value <> '',
-                                CAST(value AS int),
-                                0
-                              )
-                            FROM
-                              UNNEST(metadata)
-                            WHERE
-                              KEY = "partial"
-                        ),
-                        0
-                    ) AS isPartiallyHit
+                    ARRAY(
+                        SELECT
+                            SUM(CAST(branchHits AS INT64))
+                        FROM
+                            UNNEST(
+                                JSON_VALUE_ARRAY(
+                                    (
+                                        SELECT
+                                            value
+                                        FROM
+                                            UNNEST(metadata)
+                                        WHERE
+                                            KEY = "branchHits"
+                                    )
+                                ) 
+                            ) AS branchHits WITH OFFSET AS branchIndex
+                        GROUP BY
+                            branchIndex,
+                            branchHits
+                    ) as branchHits
                 FROM
                     `mock-table`
                 WHERE
@@ -274,21 +300,28 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                     repository = 'mock-repository'
                     
             ),
-            lineCoverage AS (
+            branchingLines AS (
                 SELECT
                     fileName,
                     lineNumber,
-                    tag,
                     SUM(hits) as hits,
-                    MIN(isPartiallyHit) as isPartiallyHit
-                FROM
-                    unnested
-                GROUP BY
+                    branchIndex,
+                    SUM(branchHit) > 0 as isBranchedLineHit
+                FROM 
+                    unnested,
+                    UNNEST(
+                        IF(
+                            ARRAY_LENGTH(branchHits) = 0,
+                            [hits],
+                            branchHits    
+                        )
+                    ) AS branchHit WITH OFFSET AS branchIndex
+                GROUP BY 
                     fileName,
                     lineNumber,
-                    tag
+                    branchIndex
             ),
-            lineCoverageWithState AS (
+            lines AS (
                 SELECT
                     fileName,
                     lineNumber,
@@ -296,13 +329,13 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                         SUM(hits) = 0,
                         "uncovered",
                         IF (
-                            MIN(isPartiallyHit) = 1,
+                            MIN(CAST(isBranchedLineHit AS INT64)) = 0,
                             "partial",
                             "covered"
                         )
                     ) as state
                 FROM
-                    lineCoverage
+                    branchingLines
                 GROUP BY
                     fileName,
                     lineNumber
@@ -321,7 +354,7 @@ class FileCoverageQueryTest extends AbstractQueryTestCase
                     2
                 ) as coveragePercentage
             FROM
-                lineCoverageWithState
+                lines
             GROUP BY
                 fileName
             ORDER BY
