+++ services/analyse/tests/Query/TotalCoverageQueryTest.php
@@ -26,22 +26,26 @@ class TotalCoverageQueryTest extends AbstractQueryTestCase
                         WHERE
                             key = "lineHits"
                     ) AS hits,
-                    IF (
-                        type = "BRANCH",
-                        (
-                            SELECT
-                              IF (
-                                value <> '',
-                                CAST(value AS int),
-                                0
-                              )
-                            FROM
-                              UNNEST(metadata)
-                            WHERE
-                              KEY = "partial"
-                        ),
-                        0
-                    ) AS isPartiallyHit
+                    ARRAY(
+                        SELECT
+                            SUM(CAST(branchHits AS INT64))
+                        FROM
+                            UNNEST(
+                                JSON_VALUE_ARRAY(
+                                    (
+                                        SELECT
+                                            value
+                                        FROM
+                                            UNNEST(metadata)
+                                        WHERE
+                                            KEY = "branchHits"
+                                    )
+                                ) 
+                            ) AS branchHits WITH OFFSET AS branchIndex
+                        GROUP BY
+                            branchIndex,
+                            branchHits
+                    ) as branchHits
                 FROM
                     `mock-table`
                 WHERE
@@ -50,21 +54,28 @@ class TotalCoverageQueryTest extends AbstractQueryTestCase
                     repository = 'mock-repository'
                     
             ),
-            lineCoverage AS (
+            branchingLines AS (
                 SELECT
                     fileName,
                     lineNumber,
-                    tag,
                     SUM(hits) as hits,
-                    MIN(isPartiallyHit) as isPartiallyHit
-                FROM
-                    unnested
-                GROUP BY
+                    branchIndex,
+                    SUM(branchHit) > 0 as isBranchedLineHit
+                FROM 
+                    unnested,
+                    UNNEST(
+                        IF(
+                            ARRAY_LENGTH(branchHits) = 0,
+                            [hits],
+                            branchHits    
+                        )
+                    ) AS branchHit WITH OFFSET AS branchIndex
+                GROUP BY 
                     fileName,
                     lineNumber,
-                    tag
+                    branchIndex
             ),
-            lineCoverageWithState AS (
+            lines AS (
                 SELECT
                     fileName,
                     lineNumber,
@@ -72,13 +83,13 @@ class TotalCoverageQueryTest extends AbstractQueryTestCase
                         SUM(hits) = 0,
                         "uncovered",
                         IF (
-                            MIN(isPartiallyHit) = 1,
+                            MIN(CAST(isBranchedLineHit AS INT64)) = 0,
                             "partial",
                             "covered"
                         )
                     ) as state
                 FROM
-                    lineCoverage
+                    branchingLines
                 GROUP BY
                     fileName,
                     lineNumber
@@ -90,7 +101,7 @@ class TotalCoverageQueryTest extends AbstractQueryTestCase
                     COALESCE(SUM(IF(state = "partial", 1, 0)), 0) as partial,
                     COALESCE(SUM(IF(state = "uncovered", 1, 0)), 0) as uncovered,
                 FROM
-                    lineCoverageWithState
+                    lines
             )
             SELECT
                 SUM(lines) as lines,
